# openapi/jobvis-v1.yaml
openapi: 3.1.0
info:
  title: Jobvis API
  version: 1.0.0
  description: Canonical HTTP contract for Jobvis v1, including the current web client surface.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /connections/capabilities:
    get:
      operationId: listConnectionCapabilities
      responses:
        '200':
          description: Provider capabilities and current availability.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ConnectionCapability' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /connections:
    get:
      operationId: listConnections
      parameters:
        - name: capability
          in: query
          schema: { type: string, enum: [mail, calendar] }
        - name: includeRevoked
          in: query
          schema: { type: boolean, default: true }
      responses:
        '200':
          description: Connections owned by the current user.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/ExternalConnection' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /connections/naver:
    post:
      operationId: connectNaverMail
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConnectNaverRequest' }
      responses:
        '201':
          description: Validated and encrypted Naver mail connection.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalConnection' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /connections/{provider}/oauth/begin:
    post:
      operationId: beginConnectionOAuth
      parameters:
        - $ref: '#/components/parameters/OAuthConnectionProvider'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BeginOAuthConnectionRequest' }
      responses:
        '200':
          description: Provider authorization URL.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BeginOAuthConnectionResponse' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /connections/{provider}/oauth/complete:
    post:
      operationId: completeConnectionOAuth
      parameters:
        - $ref: '#/components/parameters/OAuthConnectionProvider'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CompleteOAuthConnectionRequest' }
      responses:
        '201':
          description: Connected provider account.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalConnection' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '424': { $ref: '#/components/responses/FailedDependency' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /connections/{connectionId}/monitoring-consent:
    patch:
      operationId: updateConnectionMonitoringConsent
      parameters:
        - $ref: '#/components/parameters/ConnectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMonitoringConsentRequest' }
      responses:
        '200':
          description: Updated connection.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalConnection' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /connections/{connectionId}/monitoring/resume:
    post:
      operationId: resumeConnectionMonitoring
      parameters:
        - $ref: '#/components/parameters/ConnectionId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResumeMonitoringRequest' }
      responses:
        '200':
          description: Resumed connection.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ExternalConnection' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /connections/{connectionId}:
    delete:
      operationId: revokeConnection
      parameters:
        - $ref: '#/components/parameters/ConnectionId'
      responses:
        '204': { description: Connection revoked. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /import-runs:
    get:
      operationId: listImportRuns
      parameters:
        - name: page
          in: query
          schema: { type: integer, minimum: 0, default: 0 }
        - name: size
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
      responses:
        '200':
          description: Import runs ordered newest first.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportRunPage' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createImportRun
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateImportRunRequest' }
      responses:
        '202':
          description: Queued or replayed import run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportRun' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '429': { $ref: '#/components/responses/TooManyRequests' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /import-runs/{runId}:
    get:
      operationId: getImportRun
      parameters:
        - $ref: '#/components/parameters/ImportRunId'
      responses:
        '200':
          description: Import run state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportRun' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /import-runs/{runId}/cancel:
    post:
      operationId: cancelImportRun
      parameters:
        - $ref: '#/components/parameters/ImportRunId'
      responses:
        '200':
          description: Cancelled import run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportRun' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    OAuthConnectionProvider:
      name: provider
      in: path
      required: true
      schema: { type: string, enum: [gmail, outlook, google_calendar] }
    ConnectionId:
      name: connectionId
      in: path
      required: true
      schema: { type: string, format: uuid }
    ImportRunId:
      name: runId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
...
    BadRequest:
      description: Invalid request.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Unauthorized:
      description: Authentication is required.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    NotFound:
      description: Owned resource was not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    Conflict:
      description: Mutation id or optimistic version conflict.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    TooManyRequests:
      description: Request rate limit exceeded.
      headers:
        Retry-After:
          required: true
          schema: { type: integer, minimum: 1 }
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    ServiceUnavailable:
      description: A required provider or service is unavailable.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
    FailedDependency:
      description: The external provider requires another action.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/ProblemDetail' }
  schemas:
...
    ConnectionProvider:
      type: string
      enum: [gmail, outlook, naver, google_calendar]
    ConnectionCapability:
      type: object
      required:
        - provider
        - capability
        - connectionMode
        - available
        - supportsHistoricalImport
        - supportsOngoingSync
        - notes
      properties:
        provider: { $ref: '#/components/schemas/ConnectionProvider' }
        capability: { type: string, enum: [mail, calendar] }
        connectionMode: { type: string, enum: [oauth2, app_password] }
        available: { type: boolean }
        supportsHistoricalImport: { type: boolean }
        supportsOngoingSync: { type: boolean }
        notes:
          type: array
          items: { type: string }
    ExternalConnection:
      type: object
      required:
        - id
        - provider
        - capability
        - accountEmail
        - status
        - ongoingSyncConsent
        - lastSyncedAt
        - nextSyncAfter
        - lastErrorCode
        - monitoringPaused
        - tokenExpiresAt
        - grantedScopes
        - version
      properties:
        id: { type: string, format: uuid }
        provider: { $ref: '#/components/schemas/ConnectionProvider' }
        capability: { type: string, enum: [mail, calendar] }
        accountEmail: { type: string, format: email }
        status:
          type: string
          enum: [connected, reauthorization_required, error, revoked]
        ongoingSyncConsent: { type: boolean }
        lastSyncedAt: { type: [string, 'null'], format: date-time }
        nextSyncAfter: { type: [string, 'null'], format: date-time }
        lastErrorCode: { type: [string, 'null'] }
        monitoringPaused: { type: boolean }
        tokenExpiresAt: { type: [string, 'null'], format: date-time }
        grantedScopes:
          type: array
          uniqueItems: true
          items: { type: string }
        version: { type: integer, format: int64, minimum: 0 }
    ConnectNaverRequest:
      type: object
      additionalProperties: false
      required: [accountEmail, appPassword, ongoingSyncConsent]
      properties:
        accountEmail: { type: string, format: email, maxLength: 320 }
        appPassword: { type: string, minLength: 1, maxLength: 255 }
        ongoingSyncConsent: { type: boolean }
    BeginOAuthConnectionRequest:
      type: object
      additionalProperties: false
      required: [redirectUri]
      properties:
        redirectUri: { type: string, minLength: 1, maxLength: 1000 }
    BeginOAuthConnectionResponse:
      type: object
      required: [provider, authorizationUrl, expiresAt]
      properties:
        provider: { $ref: '#/components/schemas/ConnectionProvider' }
        authorizationUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
    CompleteOAuthConnectionRequest:
      type: object
      additionalProperties: false
      required: [state, code, ongoingSyncConsent]
      properties:
        state: { type: string, minLength: 1, maxLength: 500 }
        code: { type: string, minLength: 1, maxLength: 4000 }
        ongoingSyncConsent: { type: boolean }
    UpdateMonitoringConsentRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion, enabled]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
        enabled: { type: boolean }
    ResumeMonitoringRequest:
      type: object
      additionalProperties: false
      required: [expectedVersion]
      properties:
        expectedVersion: { type: integer, format: int64, minimum: 0 }
    CreateImportRunRequest:
      type: object
      additionalProperties: false
      required: [mutationId, connectionId]
      properties:
        mutationId: { type: string, format: uuid }
        connectionId: { type: string, format: uuid }
        dateFrom: { type: string, format: date }
        dateTo: { type: string, format: date }
      dependentRequired:
        dateFrom: [dateTo]
        dateTo: [dateFrom]
    ImportRun:
      type: object
      required:
        - id
        - connectionId
        - connectionVersion
        - provider
        - requestedBy
        - dateFrom
        - dateTo
        - status
        - scannedCount
        - finalizedCount
        - ignoredCount
        - duplicateCount
        - errorCode
        - startedAt
        - completedAt
        - purgeAfter
        - createdAt
      properties:
        id: { type: string, format: uuid }
        connectionId: { type: string, format: uuid }
        connectionVersion: { type: integer, format: int64, minimum: 0 }
        provider: { $ref: '#/components/schemas/ConnectionProvider' }
        requestedBy: { type: string, enum: [user, monitor] }
        dateFrom: { type: string, format: date }
        dateTo: { type: string, format: date }
        status: { type: string, enum: [queued, running, completed, failed, cancelled] }
        scannedCount: { type: integer, minimum: 0 }
        finalizedCount: { type: integer, minimum: 0 }
        ignoredCount: { type: integer, minimum: 0 }
        duplicateCount: { type: integer, minimum: 0 }
        errorCode: { type: [string, 'null'] }
        startedAt: { type: [string, 'null'], format: date-time }
        completedAt: { type: [string, 'null'], format: date-time }
        purgeAfter: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    ImportRunPage:
      type: object
      required: [items, page, size, hasNext]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ImportRun' }
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        hasNext: { type: boolean }
...
    ProblemDetail:
      type: object
      additionalProperties: true
      required: [title, status, detail]
      properties:
        type: { type: string, format: uri-reference, default: about:blank }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }