# src/main/kotlin/com/meenseek/jobvis/imports/ImportController.kt
package com.meenseek.jobvis.imports

import com.meenseek.jobvis.auth.CurrentUserProvider
import jakarta.servlet.http.HttpServletRequest
import jakarta.validation.Valid
import org.springframework.http.HttpStatus
import org.springframework.web.bind.annotation.GetMapping
import org.springframework.web.bind.annotation.PathVariable
import org.springframework.web.bind.annotation.PostMapping
import org.springframework.web.bind.annotation.RequestBody
import org.springframework.web.bind.annotation.RequestMapping
import org.springframework.web.bind.annotation.RequestParam
import org.springframework.web.bind.annotation.ResponseStatus
import org.springframework.web.bind.annotation.RestController
import java.util.UUID

@RestController
@RequestMapping("/api/v1/import-runs")
class ImportRunController(
	private val currentUserProvider: CurrentUserProvider,
	private val service: ImportRunService,
) {
	@PostMapping
	@ResponseStatus(HttpStatus.ACCEPTED)
	fun create(
		@Valid @RequestBody request: CreateImportRunRequest,
		httpRequest: HttpServletRequest,
	): ImportRunResponse = service.create(currentUserProvider.currentUserId(httpRequest), request)

	@GetMapping
	fun list(
		@RequestParam(required = false, defaultValue = "0") page: Int,
		@RequestParam(required = false, defaultValue = "50") size: Int,
		httpRequest: HttpServletRequest,
	): ImportPageResponse<ImportRunResponse> =
		service.list(currentUserProvider.currentUserId(httpRequest), page, size)

	@GetMapping("/{runId}")
	fun get(@PathVariable runId: UUID, httpRequest: HttpServletRequest): ImportRunResponse =
		service.get(currentUserProvider.currentUserId(httpRequest), runId)

	@PostMapping("/{runId}/cancel")
	fun cancel(@PathVariable runId: UUID, httpRequest: HttpServletRequest): ImportRunResponse =
		service.cancel(currentUserProvider.currentUserId(httpRequest), runId)
}